Revert "audit trail and transfoarmation work"
