+++ frontend/src/components/Style.css
@@ -477,20 +477,7 @@ button {
   width: 100%;
 }
 
-.custom-sidenav-button {
-  pointer-events: auto;
-}
-
 .custom-sidenav-button:hover {
-  background-color: var(--cds-background, #295785);
-}
-
-.cds--side-nav__link a {
-  pointer-events: auto;
-  text-decoration: none !important;
-}
-
-.cds--side-nav__link a:hover {
   background-color: var(--cds-background, #295785);
 }
 
@@ -500,10 +487,6 @@ button {
 
 .cds--side-nav__link:hover {
   background-color: var(--cds-background-hover, #072655);
-}
-
-.cds--side-nav__link {
-  pointer-events: none
 }
 
 @media screen and (max-width: 792px) {
