+++ frontend/src/components/layout/Header.js
@@ -183,7 +183,11 @@ function OEHeader(props) {
       if (level === 0 && menuItem.childMenus.length > 0) {
         return (
           <React.Fragment key={path}>
-            <span id={menuItem.menu.elementId}>
+            <span
+              onClick={(e) => {
+                setMenuItemExpanded(e, menuItem, path);
+              }}
+            >
               <SideNavMenu
                 aria-label={intl.formatMessage({
                   id: menuItem.menu.displayKey,
@@ -193,9 +197,6 @@ function OEHeader(props) {
                 })}
                 key={"menu_" + index + "_" + level}
                 defaultExpanded={menuItem.expanded}
-                onClick={(e) => {
-                  setMenuItemExpanded(e, menuItem, path);
-                }}
               >
                 {menuItem.childMenus.map((childMenuItem, index) => {
                   return generateMenuItems(
@@ -213,7 +214,6 @@ function OEHeader(props) {
         return (
           <React.Fragment key={path}>
             <SideNavMenuItem
-              id={menuItem.menu.elementId}
               href={menuItem.menu.actionURL}
               target={menuItem.menu.openInNewWindow ? "_blank" : ""}
             >
@@ -277,27 +277,19 @@ function OEHeader(props) {
   const renderSingleNavButton = (menuItem, index, level, path) => {
     const marginValue = (level - 1) * 0.5 + "rem";
     return (
-      <a
-        id={menuItem.menu.elementId}
-        href={menuItem.menu.actionURL}
-        target={menuItem.menu.openInNewWindow ? "_blank" : ""}
-        style={{ width: "100%" }}
-        rel="noreferrer"
+      <button
+        className={"custom-sidenav-button"}
+        style={{ marginLeft: marginValue }}
+        onClick={() => {
+          if (menuItem.menu.openInNewWindow) {
+            window.open(menuItem.menu.actionURL);
+          } else {
+            window.location.href = menuItem.menu.actionURL;
+          }
+        }}
       >
-        <button
-          className={"custom-sidenav-button"}
-          style={{ marginLeft: marginValue }}
-          onClick={() => {
-            if (menuItem.menu.openInNewWindow) {
-              window.open(menuItem.menu.actionURL);
-            } else {
-              window.location.href = menuItem.menu.actionURL;
-            }
-          }}
-        >
-          {renderSideNavMenuItemLabel(menuItem, level)}
-        </button>
-      </a>
+        {renderSideNavMenuItemLabel(menuItem, level)}
+      </button>
     );
   };
 
@@ -305,7 +297,6 @@ function OEHeader(props) {
     const marginValue = (level - 1) * 0.5 + "rem";
     return (
       <button
-        id={menuItem.menu.displayKey + "_dropdown"}
         className={"custom-sidenav-button"}
         style={{ marginLeft: marginValue }}
         onClick={(e) => {
@@ -322,34 +313,25 @@ function OEHeader(props) {
     const marginValue = (level - 1) * 0.5 + "rem";
     return (
       <>
-        <a
-          id={menuItem.menu.elementId}
-          href={menuItem.menu.actionURL}
-          target={menuItem.menu.openInNewWindow ? "_blank" : ""}
-          style={{ width: "100%" }}
-          rel="noreferrer"
-        >
-          <button
-            className={
-              menuItem.menu.actionURL
-                ? "custom-sidenav-button"
-                : "custom-sidenav-button-unclickable"
+        <button
+          className={
+            menuItem.menu.actionURL
+              ? "custom-sidenav-button"
+              : "custom-sidenav-button-unclickable"
+          }
+          style={{ marginLeft: marginValue }}
+          onClick={() => {
+            if (menuItem.menu.openInNewWindow) {
+              window.open(menuItem.menu.actionURL);
+            } else {
+              window.location.href = menuItem.menu.actionURL;
             }
-            style={{ marginLeft: marginValue }}
-            onClick={() => {
-              if (menuItem.menu.openInNewWindow) {
-                window.open(menuItem.menu.actionURL);
-              } else {
-                window.location.href = menuItem.menu.actionURL;
-              }
-            }}
-          >
-            {renderSideNavMenuItemLabel(menuItem, level)}
-          </button>
-        </a>
+          }}
+        >
+          {renderSideNavMenuItemLabel(menuItem, level)}
+        </button>
         {menuItem.childMenus.length > 0 && (
           <button
-            id={menuItem.menu.displayKey + "_dropdown"}
             className="custom-sidenav-button"
             onClick={(e) => {
               onClickSideNavItem(e, menuItem, path);
