+++ src/main/java/org/openelisglobal/audittrail/controller/AuditTrailReportController.java
@@ -9,7 +9,6 @@ import org.openelisglobal.audittrail.action.workers.AuditTrailItem;
 import org.openelisglobal.audittrail.action.workers.AuditTrailViewWorker;
 import org.openelisglobal.audittrail.form.AuditTrailViewForm;
 import org.openelisglobal.common.controller.BaseController;
-import org.openelisglobal.spring.util.SpringContext;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.InitBinder;
@@ -36,8 +35,7 @@ public class AuditTrailReportController extends BaseController {
 
         String accessionNumber = form.getAccessionNumberSearch();
         if (!GenericValidator.isBlankOrNull(accessionNumber)) {
-            AuditTrailViewWorker worker = SpringContext.getBean(AuditTrailViewWorker.class);
-            worker.setAccessionNumber(accessionNumber);
+            AuditTrailViewWorker worker = new AuditTrailViewWorker(accessionNumber);
             List<AuditTrailItem> items = worker.getAuditTrail();
             form.setLog(items);
             form.setAccessionNumber(accessionNumber);
