+++ src/main/java/org/openelisglobal/audittrail/controller/rest/AuditTrailReportRestController.java
@@ -4,7 +4,6 @@ import java.util.List;
 import org.openelisglobal.audittrail.action.workers.AuditTrailItem;
 import org.openelisglobal.audittrail.action.workers.AuditTrailViewWorker;
 import org.openelisglobal.audittrail.form.AuditTrailViewForm;
-import org.openelisglobal.spring.util.SpringContext;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -16,9 +15,7 @@ public class AuditTrailReportRestController {
     @GetMapping("/rest/AuditTrailReport")
     public ResponseEntity<AuditTrailViewForm> getAuditTrailReport(@RequestParam String accessionNumber) {
         AuditTrailViewForm response = new AuditTrailViewForm();
-
-        AuditTrailViewWorker worker = SpringContext.getBean(AuditTrailViewWorker.class);
-        worker.setAccessionNumber(accessionNumber);
+        AuditTrailViewWorker worker = new AuditTrailViewWorker(accessionNumber);
         List<AuditTrailItem> items = worker.getAuditTrail();
 
         if (items.size() == 0) {
