+++ src/main/java/org/openelisglobal/dataexchange/fhir/service/FhirTransformServiceImpl.java
@@ -592,8 +592,7 @@ public class FhirTransformServiceImpl implements FhirTransformService {
             strDate = strDate.replaceFirst(DateUtil.AMBIGUOUS_DATE_SEGMENT, "01");
             monthAmbiguous = true;
         }
-        Date birthDate = new SimpleDateFormat(DateUtil.getDateFormat()).parse(strDate);
-
+        Date birthDate = new SimpleDateFormat("dd/MM/yyyy").parse(strDate);
         DateType dateType = new DateType();
         if (monthAmbiguous) {
             dateType.setValue(birthDate, TemporalPrecisionEnum.YEAR);
@@ -633,7 +632,7 @@ public class FhirTransformServiceImpl implements FhirTransformService {
                 fhirPatient.setBirthDateElement(transformToDateElement(patient.getBirthDateForDisplay()));
             }
         } catch (ParseException e) {
-            LogEvent.logError("patient date unparseable " + patient.getBirthDateForDisplay(), e);
+            LogEvent.logError("patient date unparseable", e);
         }
         if (GenericValidator.isBlankOrNull(patient.getGender())) {
             fhirPatient.setGender(AdministrativeGender.UNKNOWN);
