+++ src/main/java/org/openelisglobal/program/service/ProgramSampleServiceImpl.java
@@ -14,7 +14,6 @@ public class ProgramSampleServiceImpl extends AuditableBaseObjectServiceImpl<Pro
 
     ProgramSampleServiceImpl() {
         super(ProgramSample.class);
-        this.auditTrailLog = true;
     }
 
     @Override
