+++ src/main/java/org/openelisglobal/program/service/cytology/CytologySampleServiceImpl.java
@@ -64,7 +64,6 @@ public class CytologySampleServiceImpl extends AuditableBaseObjectServiceImpl<Cy
 
     CytologySampleServiceImpl() {
         super(CytologySample.class);
-        this.auditTrailLog = true;
     }
 
     @Override
