+++ src/main/java/org/openelisglobal/program/valueholder/cytology/CytologySample.java
@@ -1,7 +1,6 @@
 package org.openelisglobal.program.valueholder.cytology;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -13,7 +12,6 @@ import javax.persistence.OneToOne;
 import javax.persistence.Table;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
-import org.apache.commons.lang3.StringUtils;
 import org.openelisglobal.program.valueholder.ProgramSample;
 import org.openelisglobal.systemuser.valueholder.SystemUser;
 
@@ -71,14 +69,6 @@ public class CytologySample extends ProgramSample {
         return technician;
     }
 
-    public String getTechnician_Audit() {
-        if (technician == null) {
-            return null;
-        } else {
-            return technician.getDisplayName();
-        }
-    }
-
     public void setTechnician(SystemUser technician) {
         this.technician = technician;
     }
@@ -95,30 +85,12 @@ public class CytologySample extends ProgramSample {
         return slides;
     }
 
-    public String getSlides_Audit() {
-        if (slides == null) {
-            return null;
-        } else {
-            return StringUtils
-                    .join(slides.stream().map(e -> "File Type: " + e.getFileType() + ", Location: " + e.getLocation())
-                            .collect(Collectors.toList()), "; ");
-        }
-    }
-
     public void setSlides(List<CytologySlide> slides) {
         this.slides = slides;
     }
 
     public SystemUser getCytoPathologist() {
         return cytoPathologist;
-    }
-
-    public String getCytoPathologist_Audit() {
-        if (cytoPathologist == null) {
-            return null;
-        } else {
-            return cytoPathologist.getDisplayName();
-        }
     }
 
     public void setCytoPathologist(SystemUser cytoPathologist) {
@@ -129,15 +101,6 @@ public class CytologySample extends ProgramSample {
         return specimenAdequacy;
     }
 
-    public String getSpecimenAdequacy_Audit() {
-        if (specimenAdequacy == null) {
-            return null;
-        } else {
-            return "Result Type: " + specimenAdequacy.getResultType() + ", Satisfaction: "
-                    + specimenAdequacy.getSatisfaction() == null ? "" : specimenAdequacy.getSatisfaction().getDisplay();
-        }
-    }
-
     public void setSpecimenAdequacy(CytologySpecimenAdequacy specimenAdequacy) {
         this.specimenAdequacy = specimenAdequacy;
     }
@@ -146,31 +109,12 @@ public class CytologySample extends ProgramSample {
         return diagnosis;
     }
 
-    public String getDiagnosis_Audit() {
-        if (diagnosis == null) {
-            return null;
-        } else {
-            return "Negative Diagnosis: " + diagnosis.getNegativeDiagnosis();
-        }
-    }
-
     public void setDiagnosis(CytologyDiagnosis diagnosis) {
         this.diagnosis = diagnosis;
     }
 
     public List<CytologyReport> getReports() {
         return reports;
-    }
-
-    public String getReports_Audit() {
-        if (reports == null) {
-            return null;
-        } else {
-            return StringUtils.join(reports.stream()
-                    .map(e -> "File Type: " + e.getFileType() + ", Report Type: "
-                            + (e.getReportType() == null ? "" : e.getReportType().getDisplay()))
-                    .collect(Collectors.toList()), "; ");
-        }
     }
 
     public void setReports(List<CytologyReport> reports) {
