+++ src/main/java/org/openelisglobal/program/valueholder/immunohistochemistry/ImmunohistochemistrySample.java
@@ -1,7 +1,6 @@
 package org.openelisglobal.program.valueholder.immunohistochemistry;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -13,7 +12,6 @@ import javax.persistence.OneToOne;
 import javax.persistence.Table;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
-import org.apache.commons.lang3.StringUtils;
 import org.openelisglobal.program.valueholder.ProgramSample;
 import org.openelisglobal.program.valueholder.pathology.PathologySample;
 import org.openelisglobal.systemuser.valueholder.SystemUser;
@@ -76,14 +74,6 @@ public class ImmunohistochemistrySample extends ProgramSample {
         return technician;
     }
 
-    public String getTechnician_Audit() {
-        if (technician == null) {
-            return null;
-        } else {
-            return technician.getDisplayName();
-        }
-    }
-
     public void setTechnician(SystemUser technician) {
         this.technician = technician;
     }
@@ -92,31 +82,12 @@ public class ImmunohistochemistrySample extends ProgramSample {
         return pathologist;
     }
 
-    public String getPathologist_Audit() {
-        if (pathologist == null) {
-            return null;
-        } else {
-            return pathologist.getDisplayName();
-        }
-    }
-
     public void setPathologist(SystemUser pathologist) {
         this.pathologist = pathologist;
     }
 
     public List<ImmunohistochemistrySampleReport> getReports() {
         return reports;
-    }
-
-    public String getReports_Audit() {
-        if (reports == null) {
-            return null;
-        } else {
-            return StringUtils.join(reports.stream()
-                    .map(e -> "File Type: " + e.getFileType() + ", Report Type: "
-                            + (e.getReportType() == null ? "" : e.getReportType().getDisplay()))
-                    .collect(Collectors.toList()), "; ");
-        }
     }
 
     public void setReports(List<ImmunohistochemistrySampleReport> reports) {
