+++ src/main/java/org/openelisglobal/program/valueholder/pathology/PathologySample.java
@@ -1,7 +1,6 @@
 package org.openelisglobal.program.valueholder.pathology;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -13,15 +12,7 @@ import javax.persistence.OneToOne;
 import javax.persistence.Table;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
-import org.apache.commons.lang3.StringUtils;
-import org.hibernate.annotations.LazyCollection;
-import org.hibernate.annotations.LazyCollectionOption;
-import org.openelisglobal.dictionary.service.DictionaryService;
 import org.openelisglobal.program.valueholder.ProgramSample;
-import org.openelisglobal.program.valueholder.pathology.PathologyConclusion.ConclusionType;
-import org.openelisglobal.program.valueholder.pathology.PathologyRequest.RequestType;
-import org.openelisglobal.program.valueholder.pathology.PathologyTechnique.TechniqueType;
-import org.openelisglobal.spring.util.SpringContext;
 import org.openelisglobal.systemuser.valueholder.SystemUser;
 
 @Entity
@@ -60,7 +51,6 @@ public class PathologySample extends ProgramSample {
 
     @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
     @JoinColumn(name = "pathology_sample_id")
-    @LazyCollection(LazyCollectionOption.FALSE)
     private List<PathologyBlock> blocks;
 
     @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
@@ -93,19 +83,6 @@ public class PathologySample extends ProgramSample {
         return requests;
     }
 
-    public String getRequests_Audit() {
-        if (requests == null) {
-            return null;
-        } else {
-            return StringUtils.join(requests.stream()
-                    .map(e -> "Status: " + (e.getStatus() == null ? "" : e.getStatus().getDisplay()) + ", Value "
-                            + (e.getType() == RequestType.DICTIONARY
-                                    ? SpringContext.getBean(DictionaryService.class).get(e.getValue()).getDisplayValue()
-                                    : e.getValue()))
-                    .collect(Collectors.toList()), "; ");
-        }
-    }
-
     public void setRequests(List<PathologyRequest> requests) {
         this.requests = requests;
     }
@@ -114,36 +91,12 @@ public class PathologySample extends ProgramSample {
         return techniques;
     }
 
-    public String getTechniques_Audit() {
-        if (techniques == null) {
-            return null;
-        } else {
-            return StringUtils.join(techniques.stream()
-                    .map(e -> "Value: " + (e.getType() == TechniqueType.DICTIONARY
-                            ? SpringContext.getBean(DictionaryService.class).get(e.getValue()).getDisplayValue()
-                            : "" + e.getValue()))
-                    .collect(Collectors.toList()), "; ");
-        }
-    }
-
     public void setTechniques(List<PathologyTechnique> techniques) {
         this.techniques = techniques;
     }
 
     public List<PathologyConclusion> getConclusions() {
         return conclusions;
-    }
-
-    public String getConclusions_Audit() {
-        if (conclusions == null) {
-            return null;
-        } else {
-            return StringUtils.join(conclusions.stream()
-                    .map(e -> "Value: " + (e.getType() == ConclusionType.DICTIONARY
-                            ? SpringContext.getBean(DictionaryService.class).get(e.getValue()).getDisplayValue()
-                            : "" + e.getValue()))
-                    .collect(Collectors.toList()), "; ");
-        }
     }
 
     public void setConclusions(List<PathologyConclusion> conclusions) {
@@ -162,28 +115,12 @@ public class PathologySample extends ProgramSample {
         return technician;
     }
 
-    public String getTechnician_Audit() {
-        if (technician == null) {
-            return null;
-        } else {
-            return technician.getDisplayName();
-        }
-    }
-
     public void setTechnician(SystemUser technician) {
         this.technician = technician;
     }
 
     public SystemUser getPathologist() {
         return pathologist;
-    }
-
-    public String getPathologist_Audit() {
-        if (pathologist == null) {
-            return null;
-        } else {
-            return pathologist.getDisplayName();
-        }
     }
 
     public void setPathologist(SystemUser pathologist) {
@@ -194,33 +131,12 @@ public class PathologySample extends ProgramSample {
         return blocks;
     }
 
-    public String getBlocks_Audit() {
-        if (blocks == null) {
-            return null;
-        } else {
-            return StringUtils.join(
-                    blocks.stream().map(e -> "Block Number: " + e.getBlockNumber() + ", Location: " + e.getLocation())
-                            .collect(Collectors.toList()),
-                    "; ");
-        }
-    }
-
     public void setBlocks(List<PathologyBlock> blocks) {
         this.blocks = blocks;
     }
 
     public List<PathologySlide> getSlides() {
         return slides;
-    }
-
-    public String getSlides_Audit() {
-        if (slides == null) {
-            return null;
-        } else {
-            return StringUtils
-                    .join(slides.stream().map(e -> "File Type: " + e.getFileType() + ", Location: " + e.getLocation())
-                            .collect(Collectors.toList()), "; ");
-        }
     }
 
     public void setSlides(List<PathologySlide> slides) {
@@ -245,17 +161,6 @@ public class PathologySample extends ProgramSample {
 
     public List<PathologyReport> getReports() {
         return reports;
-    }
-
-    public String getReports_Audit() {
-        if (reports == null) {
-            return null;
-        } else {
-            return StringUtils.join(reports.stream()
-                    .map(e -> "File Type : " + e.getFileType() + ", Report Type: "
-                            + (e.getReportType() == null ? "" : e.getReportType().getDisplay()))
-                    .collect(Collectors.toList()), "; ");
-        }
     }
 
     public void setReports(List<PathologyReport> reports) {
