+++ src/main/resources/languages/message_en.properties
@@ -155,10 +155,6 @@ audit.show.all           = All
 auditTrail.action.new    = new
 auditTrail.action.update = update
 auditTrail.order         = Order
-auditTrail.pathologySample = Pathology Sample
-auditTrail.immunohistochemistrySample = Immunohistochemistry Sample
-auditTrail.cytologySample = Cytology Sample
-auditTrail.programSample = Program Sample
 
 banner.menu.administration                        = Admin
 banner.menu.analyzer.results                      = Analyzer Results
