+++ src/main/resources/languages/message_fr.properties
@@ -170,10 +170,6 @@ audit.show.all           = Tous
 auditTrail.action.new    = neuf
 auditTrail.action.update = r\u00E9viser
 auditTrail.order         = Ordonnance
-auditTrail.pathologySample = Pathology \u00E9chantillon
-auditTrail.immunohistochemistrySample = Immunohistochemistry \u00E9chantillon
-auditTrail.cytologySample = Cytology \u00E9chantillon
-auditTrail.programSample = Program \u00E9chantillon
 
 banner.menu.administration                        = Admin
 banner.menu.analyzer.results                      = Analyzer Results
