--- src/main/resources/liquibase/2.8.x.x/sample_program_audit_trail.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<databaseChangeLog
-    xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.8.xsd">
-
-    <property name="now" value="now()" dbms="postgresql" />
-
-    <changeSet author="csteele" id="1">
-        <preConditions onFail="MARK_RAN">
-            <sqlCheck expectedResult="0"> SELECT COUNT(*) FROM clinlims.reference_tables WHERE name
-                = 'PATHOLOGY_SAMPLE' </sqlCheck>
-        </preConditions>
-        <comment>Create program sample reference tables entries</comment>
-        <insert tableName="reference_tables" schemaName="clinlims">
-            <column name="id" valueSequenceNext="reference_tables_seq" />
-            <column name="name" value="PATHOLOGY_SAMPLE" />
-            <column name="keep_history" value="Y" />
-            <column name="is_hl7_encoded" value="N" />
-            <column name="lastupdated" valueComputed="${now}" />
-        </insert>
-        <insert tableName="reference_tables" schemaName="clinlims">
-            <column name="id" valueSequenceNext="reference_tables_seq" />
-            <column name="name" value="IMMUNOHISTOCHEMISTRY_SAMPLE" />
-            <column name="keep_history" value="Y" />
-            <column name="is_hl7_encoded" value="N" />
-            <column name="lastupdated" valueComputed="${now}" />
-        </insert>
-        <insert tableName="reference_tables" schemaName="clinlims">
-            <column name="id" valueSequenceNext="reference_tables_seq" />
-            <column name="name" value="CYTOLOGY_SAMPLE" />
-            <column name="keep_history" value="Y" />
-            <column name="is_hl7_encoded" value="N" />
-            <column name="lastupdated" valueComputed="${now}" />
-        </insert>
-        <insert tableName="reference_tables" schemaName="clinlims">
-            <column name="id" valueSequenceNext="reference_tables_seq" />
-            <column name="name" value="PROGRAM_SAMPLE" />
-            <column name="keep_history" value="Y" />
-            <column name="is_hl7_encoded" value="N" />
-            <column name="lastupdated" valueComputed="${now}" />
-        </insert>
-    </changeSet>
-
-</databaseChangeLog>
